separating logical and presentational code
