+++ src/views/Package/constants.ts
@@ -34,3 +34,5 @@ export const PACKAGE_ANALYTICS = {
     ABUSE: packageEvent("Feedback", "Abuse"),
   },
 };
+
+export const API_URL_RESOURCE = "api";
